requirements.txt
openai>=1.0.0
python-telegram-bot==22.7
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
easyocr>=1.7.0
Pillow>=10.0.0
